Rename to export image.
